Refactor: Separate binding tests by category.
Categories are fairly arbitrary but loosely correspond to documented "features" and are intended to cover anywhere between a few and about a dozen separate test cases.

Mostly does not change what the tests do, or even how most are written, except for a few very old ones that were written before markup parsing was fully implemented and used manual node creation.

The purpose is to break up the original monolithic `BindingTests` that was thousands of lines and becoming very difficult to maintain or find specific test cases in.