troubleshoot: Add tracing in a few more places for view and view extensions.
The cause of the overhead is still a mystery; it shows as being in `SetHoverPath` but there are no sub-slices with `GetChildrenAt`, so either it is not being called or possibly it is getting lost or omitted from the trace due to being an enumerator method.

If the slowdown is in some other method then this should help narrow it down. And if the newly-traced `GetPathToPosition` doesn't appear in the trace then we'll know that the trace itself is not handling enumerators correctly.