Perf: Reduce `GetFloatingBounds` overhead.
In addition to inlining most of the code and eliminating LINQ from the equation, this changes the implementation to compute the bounds only on layout, i.e. when either the view itself was dirty or when it triggered a layout on a floating child.

This should, ideally, eliminate most of the overhead in methods that are meant to be passive and quick, including `ContainsPoint`.

Note, the logic has not been fully regression-tested.